requirements.txt
mysql-connector-python
prompt_toolkit